Create separate module for test_identify_repo
